fix: wrapper & useTabActive
