+++ frontend/src/App.js
@@ -3,13 +3,7 @@ import { library } from '@fortawesome/fontawesome-svg-core'
 import { fal } from '@fortawesome/pro-light-svg-icons'
 import { fab } from '@fortawesome/free-brands-svg-icons'
 
-import {
-    ErrorContextProvider,
-    OrgContextProvider,
-    UserContextProvider
-} from '@contexts';
-
-import { ActionBar, SEO, ScrollToTop, Wallet } from "@components"
+import { DashboardWrapper, ScrollToTop, SEO } from "@components"
 
 import { Dashboard, Discover, Page } from "@pages"
 
@@ -17,33 +11,8 @@ import "@style/App.css"
 
 library.add(fal, fab)
 
-const title = "The Web3 Organization Key Solution | Badger";
+const title = "No-Code Onchain Access Policies | Badger";
 const description = "Level up the access-controls of your onchain organization and enjoy the benefits of a Web3 focused key solution."
-
-const DashboardUserWrapper = ({ children }) => {
-    return (
-        <>
-            <OrgContextProvider>
-                <UserContextProvider>
-                    <ActionBar />
-                    {children}
-                </UserContextProvider>
-            </OrgContextProvider >
-        </>
-    )
-}
-
-const DashboardWrapper = ({ children }) => {
-    return (
-        <ErrorContextProvider>
-            <Wallet>
-                <DashboardUserWrapper>
-                    {children}
-                </DashboardUserWrapper>
-            </Wallet>
-        </ErrorContextProvider >
-    )
-}
 
 function App() {
     return (
@@ -55,15 +24,11 @@ function App() {
 
                 <Routes>
                     <Route exact path="/discover/" element={
-                        <DashboardWrapper>
-                            <Discover />
-                        </DashboardWrapper>
+                        <DashboardWrapper children={<Discover />} />
                     } />
 
                     <Route exact path="/dashboard/*" element={
-                        <DashboardWrapper>
-                            <Dashboard />
-                        </DashboardWrapper>
+                        <DashboardWrapper children={<Dashboard />} />
                     } />
 
                     <Route path="/*" element={<Page />} />
