+++ frontend/src/components/Dashboard/index.js
@@ -1,2 +1,3 @@
 export { Dashboard } from './Dashboard';
 export { DashboardLoader } from './DashboardLoader';
+export { DashboardWrapper } from './DashboardWrapper';
